requirements.txt
streamlit
numpy
pandas
scipy
requests
python-dotenv
